fix(spmc): 32-bit variables for 'sp_id' and 'runtime_el'
This patch fixes the length of the 'sp_id' and 'runtime_el'
variables to match the values in the SP manifest. The SP
manifest defines these properties as 32-bit values and this
change reflects the same.

+++ services/std_svc/spm/spmc/spmc.h
@@ -110,12 +110,12 @@ typedef struct secure_partition_desc {
 	/*
 	 * ID of the Secure Partition
 	 */
-	uint16_t sp_id;
+	uint32_t sp_id;
 
 	/*
 	 * Runtime EL
 	 */
-	uint16_t runtime_el;
+	uint32_t runtime_el;
 
 	/*
 	 * Mailbox tracking
